+++ eslint.config.js
@@ -1,5 +1,5 @@
 // @ts-check
-// eslint.config.js 2.0.6
+// eslint.config.js 2.0.7
 
 // This ESLint configuration is designed for a TypeScript project using ESM modules.
 
@@ -140,7 +140,6 @@ export default defineConfig([
       'no-unused-vars': 'off', // Disable base rule for unused variables and use the TypeScript-specific rule instead
       'n/no-extraneous-import': 'off', // Allow imports from devDependencies in test files
       'n/no-unpublished-import': 'off', // Allow imports from unpublished packages
-      '@typescript-eslint/no-deprecated': 'off', // Disable deprecation warnings in test files
       '@typescript-eslint/no-unused-vars': 'off', // Disable TypeScript rule for unused variables in test files
       '@typescript-eslint/no-explicit-any': 'off', // Allow 'any' type in test files
       '@typescript-eslint/no-empty-function': 'off', // Allow empty functions in test files
@@ -170,7 +169,6 @@ export default defineConfig([
       'no-unused-vars': 'off', // Disable base rule for unused variables and use the TypeScript-specific rule instead
       'n/no-extraneous-import': 'off', // Allow imports from devDependencies in test files
       'n/no-unpublished-import': 'off', // Allow imports from unpublished packages
-      '@typescript-eslint/no-deprecated': 'off', // Disable deprecation warnings in test files
       '@typescript-eslint/no-unused-vars': 'off', // Disable TypeScript rule for unused variables in test files
       '@typescript-eslint/no-explicit-any': 'off', // Allow 'any' type in test files
       '@typescript-eslint/no-empty-function': 'off', // Allow empty functions in test files